+++ internal/routing/routing.go
@@ -11,4 +11,5 @@ type Router interface {
 	Close() error
 	Resolve(ctx context.Context, key string, allowSelf bool, count int) (<-chan string, error)
 	Advertise(ctx context.Context, keys []string) error
+	HasMirrors() (bool, error)
 }