+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#117](https://github.com/XenitAB/spegel/pull/117) Update Containerd client to 1.7.
 - [#126](https://github.com/XenitAB/spegel/pull/126) Refactor registry implementation to not require separate handler.
 - [#132](https://github.com/XenitAB/spegel/pull/132) Extend tests to validate single node and mirror fallback.
+- [#133](https://github.com/XenitAB/spegel/pull/133) Use routing table size for readiness check.
 
 ### Deprecated